+++ tests/test_trajectory_reconstruction/test_orientation_methods/test_ori_method_mixin.py
@@ -77,7 +77,7 @@ class TestOrientationMethodMixin:
         fs = 204.8
 
         strides = healthy_example_stride_events["left_sensor"]
-        start, end = int(strides.iloc[:1]["start"]), int(strides.iloc[:1]["end"])
+        start, end = int(strides.iloc[0]["start"]), int(strides.iloc[0]["end"])
         data = healthy_example_imu_data["left_sensor"].iloc[start:end]
 
         test.estimate(data, sampling_rate_hz=fs)